+++ lib/Models/ItemSearchProvider.ts
@@ -1,7 +1,4 @@
 import BoundingSphere from "terriajs-cesium/Source/Core/BoundingSphere";
-import Rectangle from "terriajs-cesium/Source/Core/Rectangle";
-import DataSource from "terriajs-cesium/Source/DataSources/DataSource";
-import CameraView from "./CameraView";
 
 export type ItemSearchParameter =
   | NumericItemSearchParameter
@@ -31,10 +28,8 @@ export type ItemSearchResult = {
   id: string | number;
   idPropertyName: string;
   zoomToTarget:
-    | CameraView
-    | Rectangle
-    | DataSource
-    | { boundingSphere: BoundingSphere };
+    | BoundingSphere
+    | { latitude: number; longitude: number; featureHeight: number };
   properties: Record<string, string | number>;
 };